// token/token/src/interfaces/message-enums/token-logs.enum.ts
export enum TokenLogsEnum {
    TOKEN_CREATION_INITIATED = 'Token creation initiated',
    TOKEN_CREATION_SUCCESS = 'Token created successfully',
    TOKEN_CREATION_ERROR = 'Token creation failed',
    TOKEN_DESTRUCTION_INITIATED = 'Token destruction initiated',
    TOKEN_DESTRUCTION_ERROR = 'Token destruction error',
    TOKEN_DESTRUCTION_SUCCESS = 'Token destroyed successfully',
    TOKEN_DESTRUCTION_NOT_FOUND = 'Token not found for destruction',
    TOKEN_DECODE_INITIATED = 'Token decode initiated',
    TOKEN_DECODE_SUCCESS = 'Token decoded successfully',
    TOKEN_DECODE_NOT_FOUND = 'Token not found for decoding',
    TOKEN_DECODE_ERROR = 'Token decoding failed',
}